Version 0.8.6 "Push It To The Limit 🤟"
Avoid double GPU upload for overlay manager when resize/refresh.
Update OBJ reading for compute normal+tagent space.
Update transferManager to clear image color.
Add the "Color projection map" to lights.
Add AnimatedCubemapTexture.
Add caustic generator.
Update all resources loading to async mode.
Update parametric geometry generation.
Re-enable post-processing.
Add several lens effects for color changes, HSV, film grain, vignettage, ...
Add Bloom, SSR, SSAO, Depth-of-field and tone-mapping effects.

+++ src/Audio/Recorder.cpp
@@ -312,7 +312,7 @@ namespace EmEn::Audio
 					{
 						const auto bytes = renderBuf.size() * sizeof(int16_t);
 						m_outputFileStream.write(reinterpret_cast< const char * >(renderBuf.data()), static_cast< std::streamsize >(bytes));
-						m_streamByteCount += bytes;
+						m_streamByteCount += static_cast< uint32_t >(bytes);
 					}
 				}
